requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0

# Vision + RAG
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
openai>=1.0.0
pypdf>=4.0.0

# Logging & Monitoring
python-json-logger>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
